+++ src/select/docs/examples/editableoptions/filters.js
@@ -1,6 +1,0 @@
-angular.module('selectDemo')
-    .factory('mySaveLastQuery', function() {
-        return function(removedValue, lastQuery) {
-            return removedValue.name;
-        };
-    });